lib: sbi_domain: Support Zcheripurecap ABI in struct sbi_domain
Change the next_addr and next_arg1 in struct sbi_domain to uintptr_t to allow it
storing capability with boot address for next boot domain and pointer points to
DTB (arg1) for Zcheripurecap ABI.

+++ include/sbi/sbi_domain.h
@@ -187,9 +187,9 @@ struct sbi_domain {
 	/** HART id of the HART booting this domain */
 	u32 boot_hartid;
 	/** Arg1 (or 'a1' register) of next booting stage for this domain */
-	unsigned long next_arg1;
+	uintptr_t next_arg1;
 	/** Address of next booting stage for this domain */
-	unsigned long next_addr;
+	uintptr_t next_addr;
 	/** Privilege mode of next booting stage for this domain */
 	unsigned long next_mode;
 	/** Is domain allowed to reset the system */
